delete proximity_forest_utils
